Fix winRates, and top8s count
